move url param loading logic to the appshell - ui state loaded from db and then search params override ui state - fixes issue where urls would not set ui state properly
